change lib name to scl
