+++ scl/compressors/shannon_fano_elias_coder.py
@@ -8,13 +8,13 @@ which motivates Arithmetic coding
 
 """
 
-from stanford_compress.utils.bitarray_utils import float_to_bitarrays, BitArray
-from stanford_compress.utils.test_utils import get_random_data_block, try_lossless_compression
-from stanford_compress.compressors.prefix_free_compressors import (
+from scl.utils.bitarray_utils import float_to_bitarrays, BitArray
+from scl.utils.test_utils import get_random_data_block, try_lossless_compression
+from scl.compressors.prefix_free_compressors import (
     PrefixFreeEncoder,
     PrefixFreeDecoder,
 )
-from stanford_compress.core.prob_dist import ProbabilityDist
+from scl.core.prob_dist import ProbabilityDist
 import math
 import numpy as np
