+++ scl/core/data_stream.py
@@ -2,7 +2,7 @@ import abc
 import tempfile
 import os
 import typing
-from stanford_compress.core.data_block import DataBlock
+from scl.core.data_block import DataBlock
 
 Symbol = typing.Any
