+++ scl/external_compressors/zstd_external.py
@@ -1,57 +1,70 @@
-"""External compressors gzip for testing/benchmarking purposes.
+"""External compressors zstd for testing/benchmarking purposes.
 
-Zlib format is described in https://datatracker.ietf.org/doc/html/rfc1950.
-It mostly relies on Deflate which is described in https://datatracker.ietf.org/doc/html/rfc1951.
-Zlib library website: https://www.zlib.net/. 
-Also see https://aws.amazon.com/blogs/opensource/improving-zlib-cloudflare-and-comparing-performance-with-other-zlib-forks/
-for information on more efficient implementations.
+Zstd is a modern LZ77 based compressor, using ANS for encoding match lengths and distances, and
+Huffman coding for literals.
 
-We use the python zlib module (part of standard library) which internally calls the C library.
+Zstd format is described in https://github.com/facebook/zstd/blob/dev/doc/zstd_compression_format.md.
+
+Zstd website: https://github.com/facebook/zstd/.
+
+We use the python zstandard module which internally calls the C library.
 """
 
 import os
 import tempfile
-from stanford_compress.core.data_block import DataBlock
-from stanford_compress.core.data_encoder_decoder import DataDecoder, DataEncoder
-from stanford_compress.core.data_stream import TextFileDataStream, Uint8FileDataStream
-from stanford_compress.core.encoded_stream import EncodedBlockReader, EncodedBlockWriter
-from stanford_compress.core.prob_dist import ProbabilityDist
-from stanford_compress.utils.bitarray_utils import BitArray, bitarray_to_uint, uint_to_bitarray
-from stanford_compress.utils.test_utils import (
+from scl.core.data_block import DataBlock
+from scl.core.data_encoder_decoder import DataDecoder, DataEncoder
+from scl.core.data_stream import Uint8FileDataStream
+from scl.core.encoded_stream import EncodedBlockReader, EncodedBlockWriter
+from scl.core.prob_dist import ProbabilityDist
+from scl.utils.bitarray_utils import BitArray, bitarray_to_uint, uint_to_bitarray
+from scl.utils.test_utils import (
     create_random_binary_file,
     try_file_lossless_compression,
     try_lossless_compression,
 )
-import zlib
+import zstandard
 
 
-class ZlibExternalEncoder(DataEncoder):
+class ZstdExternalEncoder(DataEncoder):
     def __init__(self, level=6):
         self.level = level
         # state stays alive across blocks so we can benefit
-        self.zlib_context = zlib.compressobj(level=self.level)
+        self.zstd_obj = ZstdExternalEncoder.create_zstd_compressor_object(level=self.level)
         self.block_size_num_bits = 32  # num bits used to encode the block size at start
 
     def reset(self):
-        # start new zlib context
-        self.zlib_context = zlib.compressobj(level=self.level)
+        # start new zstd context
+        self.zstd_obj = ZstdExternalEncoder.create_zstd_compressor_object(level=self.level)
+
+    @staticmethod
+    def create_zstd_compressor_object(level: int):
+        """Create zstd compressor object with similar API as python zlib object.
+
+        Args:
+            level (int): Zstd level
+        """
+        zstd_cctx = zstandard.ZstdCompressor(level=level)
+        zstd_obj = zstd_cctx.compressobj()
+        return zstd_obj
 
     def encode_block(self, data_block: DataBlock):
         raw_bytes = bytes(data_block.data_list)
 
-        # flush below with Z_SYNC_FLUSH that ensures decompress is able to decompress the
+        # flush below with COMPRESSOBJ_FLUSH_BLOCK that ensures decompress is able to decompress the
         # data till now. Note that this still utilizes this block for finding matches when
-        # we are compressing the next block (as opposed to Z_FULL_FLUSH that resets the state).
-        # See https://www.zlib.net/manual.html for more information
-        compressed_bytes = self.zlib_context.compress(raw_bytes) + self.zlib_context.flush(
-            zlib.Z_SYNC_FLUSH
+        # we are compressing the next block (as opposed to COMPRESSOBJ_FLUSH_FINISH that resets the state).
+        # See https://python-zstandard.readthedocs.io/en/latest/compressor.html#zstandard.ZstdCompressionObj.flush
+        # for more information
+        compressed_bytes = self.zstd_obj.compress(raw_bytes) + self.zstd_obj.flush(
+            zstandard.COMPRESSOBJ_FLUSH_BLOCK
         )
 
         # FIXME: might be inefficient to convert to BitArray since it will be later be
         # converted back to bytes when writing to file
         # FIXME: also, should we worry about endianness?
 
-        # at start write the compressed size because zlib decoder expects to get complete blocks
+        # at start write the compressed size because zstd decoder expects to get complete blocks
         # and cannot determine when we want to end
         compressed_bitarray = BitArray(
             uint_to_bitarray(len(compressed_bytes) * 8, bit_width=self.block_size_num_bits)
@@ -73,26 +86,33 @@ class ZlibExternalEncoder(DataEncoder):
                 self.encode(fds, block_size=block_size, encode_writer=writer)
 
 
-class ZlibExternalDecoder(DataDecoder):
+class ZstdExternalDecoder(DataDecoder):
     def __init__(self, level=6):
         self.level = level
-        self.zlib_context = zlib.decompressobj()
+        self.zstd_obj = ZstdExternalDecoder.create_zstd_decompressor_object()
         self.block_size_num_bits = 32  # num used to encode the block size at start
 
     def reset(self):
-        self.zlib_context = zlib.decompressobj()
+        self.zstd_obj = ZstdExternalDecoder.create_zstd_decompressor_object()
+
+    @staticmethod
+    def create_zstd_decompressor_object():
+        """Create zstd compressor object with similar API as python zlib object."""
+        zstd_dctx = zstandard.ZstdDecompressor()
+        zstd_obj = zstd_dctx.decompressobj()
+        return zstd_obj
 
     def decode_block(self, compressed_bitarray: BitArray):
-        # first read the size of the gzipped block(s)
-        zlib_size = bitarray_to_uint(compressed_bitarray[: self.block_size_num_bits])
-        zlib_compressed = compressed_bitarray[
-            self.block_size_num_bits : self.block_size_num_bits + zlib_size
+        # first read the size of the zstd block(s)
+        zstd_size = bitarray_to_uint(compressed_bitarray[: self.block_size_num_bits])
+        zstd_compressed = compressed_bitarray[
+            self.block_size_num_bits : self.block_size_num_bits + zstd_size
         ]
 
-        compressed_bytes = zlib_compressed.tobytes()
+        compressed_bytes = zstd_compressed.tobytes()
         return (
-            DataBlock(list(self.zlib_context.decompress(compressed_bytes))),
-            self.block_size_num_bits + zlib_size,
+            DataBlock(list(self.zstd_obj.decompress(compressed_bytes))),
+            self.block_size_num_bits + zstd_size,
         )
 
     def decode_file(self, encoded_file_path: str, output_file_path: str):
@@ -109,9 +129,9 @@ class ZlibExternalDecoder(DataDecoder):
                 self.decode(reader, fds)
 
 
-def test_zlib_encode_decode():
-    encoder = ZlibExternalEncoder()
-    decoder = ZlibExternalDecoder()
+def test_zstd_encode_decode():
+    encoder = ZstdExternalEncoder()
+    decoder = ZstdExternalDecoder()
 
     # create some sample data consisting of bytes
     data_list = [0, 0, 1, 3, 4, 100, 255, 123, 234, 42, 186]
@@ -123,17 +143,17 @@ def test_zlib_encode_decode():
     assert is_lossless
 
 
-def test_zlib_file_encode_decode():
-    """full test for ZlibExternalEncoder and ZlibExternalDecoder
+def test_zstd_file_encode_decode():
+    """full test for ZstdExternalEncoder and ZstdExternalDecoder
 
     - create a sample file
-    - encode the file using ZlibExternalEncoder
+    - encode the file using ZstdExternalEncoder
     - perform decoding and check if the compression was lossless
 
     """
     # define encoder, decoder
-    encoder = ZlibExternalEncoder()
-    decoder = ZlibExternalDecoder()
+    encoder = ZstdExternalEncoder()
+    decoder = ZstdExternalDecoder()
 
     with tempfile.TemporaryDirectory() as tmpdirname:
         # create a file with some random data
