+++ scl/utils/bitarray_utils.py
